// frontend/src/modules/schools/schemas/createSchoolSchema.ts
import { z } from "zod";

export const createSchoolSchema = z.object({
  name: z.string().min(2, "School name must be at least 2 characters"),
  phone: z.string().min(10, "Phone number is required"),
  logo: z.string().url("Invalid URL").optional().nullable().or(z.literal("")),
});

export type CreateSchoolFormValues = z.infer<typeof createSchoolSchema>;